a assistant/requirements.txt
cohere==5.16.3
python-telegram-bot==22.3
